+++ commands/deploy.ts
@@ -6,7 +6,7 @@ import { QcloudCommand } from '../lib'
 
 export class QcloudDeploy extends QcloudCommand {
     service
-    templates
+    template
 
     constructor(serverless, options) {
         super(serverless, options)
@@ -14,7 +14,7 @@ export class QcloudDeploy extends QcloudCommand {
         this.hooks = {
             'before:deploy:deploy': async () => {
                 await this.validate()
-                await this.loadTemplates()
+                await this.loadTemplate()
             },
 
             'deploy:deploy': async () => {
@@ -23,18 +23,28 @@ export class QcloudDeploy extends QcloudCommand {
                 await this.setupFunctions()
                 await this.setupEvents()
             },
+
+            'deploy:function:initialize': async () => {
+                await this.checkIfFunctionExists()
+            },
+
+            'deploy:function:packageFunction': async () => {
+                await this.serverless.pluginManager.spawn('package:function')
+            },
+
+            'deploy:function:deploy': async () => {
+                await this.deployFunction()
+            },
         }
     }
 
-    async loadTemplates() {
-        const { serverless: { config, utils } } = this
-        const createFilePath = path.join(config.servicePath, '.serverless', 'configuration-template-create.json')
-        const updateFilePath = path.join(config.servicePath, '.serverless', 'configuration-template-update.json')
+    /* deploy */
 
-        this.templates = {
-            create: utils.readFileSync(createFilePath),
-            update: utils.readFileSync(updateFilePath),
-        }
+    async loadTemplate() {
+        const { serverless: { config, utils } } = this
+        const TemplatePath = path.join(config.servicePath, '.serverless', 'configuration-template.json')
+
+        this.template = utils.readFileSync(TemplatePath)
     }
 
     async setupService() {
@@ -43,8 +53,8 @@ export class QcloudDeploy extends QcloudCommand {
     }
 
     async createBucketIfNotExists() {
-        const { templates, provider, serverless: { cli } } = this
-        const bucket = templates.create.Resources.DeploymentBucket
+        const { template, provider, serverless: { cli } } = this
+        const bucket = template.Resources.DeploymentBucket
         const cosBucket = provider.getCOSBucket(bucket)
         const api = provider.sdk.cos
 
@@ -65,8 +75,8 @@ export class QcloudDeploy extends QcloudCommand {
     }
 
     async createAPIGatewayIfNotExists() {
-        const { templates, provider, serverless: { cli } } = this
-        const { APIGateway } = templates.create.Resources
+        const { template, provider, serverless: { cli } } = this
+        const { APIGateway } = template.Resources
         const api = provider.sdk.apigateway.setRegion(APIGateway.Region)
 
         const services = await api.describeServicesStatus({
@@ -87,13 +97,13 @@ export class QcloudDeploy extends QcloudCommand {
 
             const res = await api.createService(APIGateway)
 
-            _.assign(templates.update.Resources.APIGateway, res)
+            _.assign(template.Resources.APIGateway, res)
         } else {
             cli.log(`API Gateway service "${APIGateway.serviceName}" already exists`)
 
             const ag = matchedServices[0]
 
-            _.assign(templates.update.Resources.APIGateway, ag)
+            _.assign(template.Resources.APIGateway, ag)
 
             if (ag.serviceDesc === APIGateway.serviceDesc && ag.protocol === APIGateway.protocol) return
 
@@ -104,50 +114,72 @@ export class QcloudDeploy extends QcloudCommand {
     }
 
     async uploadArtifacts() {
-        const { templates, provider, serverless: { cli } } = this
+        const { template, provider, serverless: { cli } } = this
         const api = provider.sdk.cos
-        const bucket = templates.update.Resources.DeploymentBucket
+        const bucket = template.Resources.DeploymentBucket
         const cosBucket = provider.getCOSBucket(bucket)
 
-        cosBucket.Body = fs.createReadStream(cosBucket.Body)
+        // Shared package
+        if (cosBucket.Body) {
+            cosBucket.Body = fs.createReadStream(cosBucket.Body)
 
-        cli.log(`Uploading "${bucket.Key}" to OSS bucket "${bucket.Bucket}"...`)
+            cli.log(`Uploading "${bucket.Key}" to OSS bucket "${bucket.Bucket}"...`)
 
-        await api.putObjectAsync(cosBucket)
+            return api.putObjectAsync(cosBucket)
+        }
+
+        // There are 2 ways reach here: 1. package individually true. 2. serverless-webpack
+
+        const codes = _.uniq<string>(template.Resources.CloudFunctions.map(func => path.resolve(func.code)))
+
+        return Promise.all(codes.map(code => {
+            const zipName = path.basename(code)
+            const copyBucket = Object.assign({}, cosBucket, {
+                Key: `${cosBucket.Key}${zipName}`,
+                Body: fs.createReadStream(code),
+                ContentLength: fs.statSync(code).size,
+            })
+
+            cli.log(`Uploading "${zipName}" to OSS bucket "${bucket.Bucket}"...`)
+
+            return api.putObjectAsync(copyBucket)
+        }))
     }
 
     async setupFunctions() {
-        const { provider, templates, serverless: { cli } } = this
+        const { provider, template, serverless: { cli } } = this
         const sdk = provider.sdk.scf
-        const bucket = templates.update.Resources.DeploymentBucket
-        const functions = templates.update.Resources.CloudFunctions || []
-        const cosBucket = provider.getCOSBucket(bucket)
-
-        const codeObject = {
-            cosBucketName: bucket.Bucket, // TODO: scf api append appid?
-            cosObjectName: cosBucket.Key,
-        }
+        const bucket = template.Resources.DeploymentBucket
+        const functions = template.Resources.CloudFunctions || []
 
         return Promise.all(functions.map(func => {
+            const codeObject = {
+                cosBucketName: bucket.Bucket, // TODO: scf api append appid?
+                cosObjectName: bucket.Key,
+            }
+
+            if (bucket.Key.endsWith('/')) {
+                codeObject.cosObjectName += path.basename(func.code)
+            }
+
             return sdk.requestAsync(_.assign(
+                { Action: 'GetFunction' },
                 _.pick(func, 'Region', 'functionName'),
-                { Action: 'GetFunction' }
             ))
                 .catch(err => {
                     cli.log('ERROR: Qcloud SCF GetFunction fail!')
                     throw err.error
                 })
                 .then(res => {
-                    if (res.code !== 0) {
+                    if (res.code > 0) {
                         cli.log(`Creating function "${func.functionName}"...`)
                         return sdk.requestAsync(_.assign(
-                            {},
-                            func,
                             {
                                 Action: 'CreateFunction',
                                 Region: func.Region,
                                 codeObject,
-                            }
+                            },
+                            _.omit(func, 'code'),
                         ))
                             .catch(err => {
                                 cli.log('ERROR: Qcloud SCF CreateFunction fail!')
@@ -165,13 +197,12 @@ export class QcloudDeploy extends QcloudCommand {
                         cli.log(`Updating function "${func.functionName}"...`)
 
                         return sdk.requestAsync(_.assign(
-                            {},
-                            func,
                             {
                                 Action: 'UpdateFunction',
                                 codeType: 'Cos',
                                 codeObject,
-                            }
+                            },
+                            func,
                         ))
                             .catch(err => {
                                 cli.log('ERROR: Qcloud SCF UpdateFunction fail!')
@@ -186,9 +217,9 @@ export class QcloudDeploy extends QcloudCommand {
     triggers
 
     async setupEvents() {
-        const { templates } = this
+        const { template } = this
 
-        this.apis = templates.update.Resources.APIGatewayApis
+        this.apis = template.Resources.APIGatewayApis
         this.triggers = []
 
         await this.createApisIfNeeded()
@@ -198,8 +229,8 @@ export class QcloudDeploy extends QcloudCommand {
     async createApisIfNeeded() {
         if (!this.apis.length) return
 
-        const { provider, templates, serverless: { cli } } = this
-        const { Resources } = templates.update
+        const { provider, template, serverless: { cli } } = this
+        const { Resources } = template
         const { apigateway } = provider.sdk
         const { serviceId } = Resources.APIGateway
 
@@ -235,8 +266,8 @@ export class QcloudDeploy extends QcloudCommand {
     }
 
     async releaseAPIGateway() {
-        const { provider, options, service, templates, serverless: { cli } } = this
-        const { APIGateway } = templates.update.Resources
+        const { provider, options, service, template, serverless: { cli } } = this
+        const { APIGateway } = template.Resources
         const { apigateway } = provider.sdk
         const stage = _.get(options, 'stage')
             || _.get(service, 'provider.stage')
@@ -257,8 +288,88 @@ export class QcloudDeploy extends QcloudCommand {
             releaseDesc: 'Release by serverless'
         })
 
-        templates.update.Resources.APIGatewayRelease = _.assign({}, res, {
+        template.Resources.APIGatewayRelease = _.assign({}, res, {
             environmentName: envMap[stage],
         })
+    }
+
+    /* deploy function */
+
+    async checkIfFunctionExists() {
+        const { options, provider, serverless: { service, cli } } = this
+        const sdk = provider.sdk.scf
+        const funcObj = service.getFunction(options.function)
+
+        const res = await sdk.requestAsync({
+            Action: 'GetFunction',
+            Region: provider.region,
+            functionName: funcObj.name,
+        })
+
+        if (res.code !== 0) {
+            const errorMessage = [
+                `The function "${options.function}" you want to update is not yet deployed.`,
+                ' Please run "serverless deploy" to deploy your service.',
+                ' After that you can redeploy your services functions with the',
+                ' "serverless deploy function" command.',
+            ].join('')
+            throw new Error(errorMessage)
+        }
+    }
+
+    async deployFunction() {
+        const { options, provider, serverless: { service, config, cli } } = this
+        const sdk = provider.sdk.scf
+        const funcObj = service.getFunction(options.function)
+        const artifactFileName = `${options.function}.zip`
+
+        let artifactFilePath
+        if (_.has(funcObj, ['package', 'artifact'])) {
+            artifactFilePath = funcObj.package.artifact
+        } else {
+            const packagePath = options.package ||
+                service.package.path ||
+                path.join(config.servicePath || '.', '.serverless')
+
+            artifactFilePath = service.package.artifact ||
+                path.join(packagePath, artifactFileName)
+        }
+
+        const stats = fs.statSync(artifactFilePath)
+
+        if (stats.size > 10 * 1000 * 1000) {
+            const errorMessage = [
+                `The function zip "${artifactFileName}" you want to update is big than 10M.`,
+                ' Qcloud not allow upload directly.',
+                ' Consider reduce your package size by serverless-webpack.',
+            ].join('')
+            throw new Error(errorMessage)
+        }
+
+        cli.log(`Uploading function: ${options.function} (${Math.round(stats.size / 1000) / 1000}M)...`)
+
+        const params = {
+            Action: 'UpdateFunction',
+            Region: provider.region,
+            functionName: funcObj.name,
+            code: `@${fs.readFileSync(artifactFilePath).toString('base64')}`,
+            handler: funcObj.handler,
+            description: funcObj.description,
+            memorySize: funcObj.memorySize || _.get(service, 'provider.memorySize'),
+            timeout: funcObj.timeout || _.get(service, 'provider.timeout'),
+            runtime: _.capitalize(funcObj.runtime || _.get(service, 'provider.runtime')),
+            codeType: 'Zipfile',
+        }
+
+        const res = await sdk.requestAsync(params)
+
+        if (res.code > 0) {
+            const error = new Error(res.message)
+            error.name = res.codeDesc
+
+            throw error
+        }
+
+        cli.log(`Successfully deployed function: ${options.function}`)
     }
 }